docker/init_flightDB/requirements.txt
Flask==1.1.1
Flask-Cors==3.0.8
Flask-SQLAlchemy==2.4.1
mysql-connector-python==8.0.18
requests==2.22.0
psycopg2==2.8.4
python-dateutil==2.8.1